Refactor UrlFileHandler and AzureFileHandler to share common http download.
